Only do JS timeout when it makes sense.
The timeout code relies on redirecting the user to a timeout page that
is currently namespaced under a policy, i.e. "/:policy/timeout". To be
able to do that, we need to know the policy, either based on the :policy
routing parameters, or based on the current_claim in the session. There
is one situation where we will have neither, and that's when we're
in maintenance mode: because of the way the catch-all routing for
maintenance mode, we don't get a :policy param. Besides which it
doesn't really make sense to auto-timeout and redirect the user when
we're in maintenance mode as the route we'll redirect them to will be
rendered as the maintenance page!