Make maintenance mode policy-aware
When we introduce Maths & Physics, we're going to need to be able to
have separate maintenance modes for each policy, otherwise it going to
be more difficult to do things like end the private beta for Maths &
Physics without also taking down the Student Loans service.

This change moves the maintenance mode routing to only kick in under
the /:policy path, which a) means we have the `:policy` param to show
the appropriate policy name and generate the appropriate policy-aware
support links in the footer, but also b) means that the routes that will
be shared between the two policies (namely those for school search and
verify) will not trigger maintenance mode and thus allow the other
policy to continue to function.