Remove default policy from routing
A lot of the previous Maths & Physics preparation PRs were leading to
this moment where we could remove the default policy in the routing and
be explicit about the current active `:policy` in the app code. Now it
will be possible to introduce a second policy routing name without
breaking the existing views and controller code.

Note that a bunch of the specs are coupled to the student-loans policy
when they don't actually need to be. Future work will focus on
re-organising the feature and request specs to make it more obvious
which ones are generic service-level specs, and which ones are specific
to the policy.