Prevent service operator from approving an inconsistent claim
This implements the work hinted at in 2d7bf9b. If a service operator
views a claim for which there is a claim from the same claimant which
has been approved but not yet payrolled, and which has inconsistent
personal details which would prevent the two claims from being combined
into a single payment, then inform the service operator that this has
happened and prevent them from approving the claim.

This means that it is now unlikely that we will ever violate the
validations introduced on Payment in 2d7bf9b, but we keep them as an
additional safety net.

We still need to decide how a service operator can resolve this situation
of an inconsistent claim. For example, one option would be to contact
the claimant and ask them whether they are happy to be paid in two
different months or to ask them to re-submit their claim. If we wanted
to enable this sort of conversation, we may need to think about exposing
a subset of the claimant's bank details to the service operator (in a
privacy-preserving manner).

The implementation and user interface follow the same patterns that we
use for the "payroll gender missing" situation.

My initial implementation of Claim::ClaimsPreventingPaymentFinder tried
to let the database find the claims with inconsistent attributes. I
first tried a simple WHERE, interpolating the attribute values from the
source claim using `read_attribute`, but this fell apart when
encountering the enum student_loan_plan column, which tried to
interpolate a string instead of the integer value that the database
understands. So then I implemented it as a self-join on the claims
table, letting the database take care of the comparison between columns.
This worked but it was quite hard to read. So, Stu pointed out that in
reality, the number of payrollable claims from the same claimant is
going to be very low, so we can just do the inconsistent attributes
check programmatically.