Add a claim overview section for service operators
As part of the work to improve the claim checking process for service
operators we now have an overview of the claim with pertinent
information.

The check page will eventually have a task list of checks for the
service operator to carry out in order to approve the claim.

After speaking with Mark he'd like to keep the current 'show' claim
page as is for now.

So that we can iterate on the task list whilst preserving the current
claim show page, I've added a 'View claim check' link. This will
eventually be removed